+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ![GitHub release](https://img.shields.io/github/downloads/dovrosenberg/fvtt-campaign-builder/v1.9.0/module.zip)
 
 - You can now create color schemes to apply to custom text blocks in story webs.
+- You can now export a story web as a .PNG file from the directory context menu.
 - Fixed issue with drag/drop of rows not working properly on very long tables.
 
 ## 1.8.6 - Assorted bug fixes and improvements
@@ -13,7 +14,7 @@
 - You can now change the style of story web connections between custom text and dangers.
 - When copying text to the clipboard from a table text box or an editor, UUIDs will no longer be dropped and instead will be copied as the name of the document/entity.
 - Similarly, copying text with bullet or numbered lists will now include the bullets/numbers in the copied text when you paste it.
-- Connecting two entries in a storyweb now lets you add a relationship label.
+- Connecting two entries in a story web now lets you add a relationship label.
 - When adding a danger to a story web, now uses a dropdown instead of a typeahead, so you don't have to remember what it's named.
 - Removed the journals column from the lore/entries tab to simplify the interface.  Journal references that were previously in the journals column will be added to the description field when the module is updated.